+++ examples/simple-gulp/Gulpfile.js
@@ -1,0 +1,5 @@
+var gulp = require('gulp');
+
+gulp.task('default', function () {
+    console.log('hello world');
+});