+++ src/main/groovy/com/moowork/gradle/grunt/GruntTask.groovy
@@ -6,42 +6,36 @@ import org.gradle.api.GradleException
 class GruntTask
     extends NodeTask
 {
-    private final static String GRUNT_SCRIPT = 'node_modules/grunt-cli/bin/grunt';
+    private final static String GRUNT_SCRIPT = 'node_modules/grunt-cli/bin/grunt'
 
-    public GruntTask()
+    GruntTask()
     {
-        this.group = 'Grunt';
+        this.group = 'Grunt'
     }
 
     @Override
     void exec()
     {
-        def localGrunt = this.project.file( new File( this.project.node.nodeModulesDir, GRUNT_SCRIPT ) )
+        def gruntFile = "${this.project.grunt.workDir}/${this.project.grunt.gruntFile}"
+        def localGrunt = this.project.file( new File( (File) this.project.node.nodeModulesDir, GRUNT_SCRIPT ) )
         if ( !localGrunt.isFile() )
         {
             throw new GradleException(
                 "Grunt-CLI not installed in node_modules, please first run 'gradle ${GruntPlugin.GRUNT_INSTALL_NAME}'" )
         }
 
-        def addArg = { List<?> original, def newArg ->
-            def tempArgs = []
-            tempArgs.addAll( original )
-            tempArgs.add( newArg )
-            return tempArgs
-        }
-
         // On multi project scenario, need to specify Gruntfile.js location
-        setArgs( addArg( args as List, "--gruntfile ${this.project.grunt.workDir}/Gruntfile.js" ) )
+        addArgs( '--gruntfile', gruntFile )
 
         // If colors are disabled, add --no-color to args
         if ( !this.project.grunt.colors )
         {
-            setArgs( addArg( args as List, '--no-color' ) )
+            addArgs( '--no-color' )
         }
 
         // If output should be buffered (useful when running in parallel)
         // set standardOutput of ExecRunner to a buffer
-        def bufferedOutput
+        ByteArrayOutputStream bufferedOutput
         if ( this.project.grunt.bufferOutput )
         {
             bufferedOutput = new ByteArrayOutputStream()
@@ -61,9 +55,9 @@ class GruntTask
         finally
         {
             // If we were buffering output, print it
-            if ( this.project.grunt.bufferOutput )
+            if ( this.project.grunt.bufferOutput && ( bufferedOutput != null ) )
             {
-                println "Output from ${this.project.grunt.workDir}/Gruntfile.js"
+                println "Output from ${gruntFile}"
                 println bufferedOutput.toString()
             }
         }