+++ src/main/groovy/com/moowork/gradle/node/NodePlugin.groovy
@@ -3,7 +3,8 @@ package com.moowork.gradle.node
 import com.moowork.gradle.node.npm.NpmInstallTask
 import com.moowork.gradle.node.npm.NpmSetupTask
 import com.moowork.gradle.node.npm.NpmTask
-import com.moowork.gradle.node.task.*
+import com.moowork.gradle.node.task.NodeTask
+import com.moowork.gradle.node.task.SetupTask
 import com.moowork.gradle.node.variant.VariantBuilder
 import com.moowork.gradle.node.yarn.YarnInstallTask
 import com.moowork.gradle.node.yarn.YarnSetupTask
@@ -28,7 +29,7 @@ class NodePlugin
     void apply( final Project project )
     {
         this.project = project
-        this.config = this.project.extensions.create( NodeExtension.NAME, NodeExtension, this.project )
+        this.config = NodeExtension.create( this.project )
 
         addGlobalTypes()
         addTasks()
@@ -67,15 +68,25 @@ class NodePlugin
     private void addNpmRule()
     {
         // note this rule also makes it possible to specify e.g. "dependsOn npm_install"
+        def workingDir
+        this.project.afterEvaluate {
+            workingDir = this.project.node.nodeModulesDir
+        }
         project.getTasks().addRule( 'Pattern: "npm_<command>": Executes an NPM command.' ) { String taskName ->
             if ( taskName.startsWith( "npm_" ) )
             {
                 NpmTask npmTask = project.getTasks().create( taskName, NpmTask.class )
+                if ( workingDir )
+                {
+                    npmTask.afterEvaluate( workingDir )
+                }
+
                 String[] tokens = taskName.split( '_' ).tail() // all except first
                 npmTask.npmCommand = tokens
 
-                if (tokens.head().equalsIgnoreCase("run")) {
-                    npmTask.dependsOn(NpmInstallTask.NAME)
+                if ( tokens.head().equalsIgnoreCase( "run" ) )
+                {
+                    npmTask.dependsOn( NpmInstallTask.NAME )
                 }
 
                 return npmTask
@@ -93,8 +104,9 @@ class NodePlugin
                 String[] tokens = taskName.split( '_' ).tail() // all except first
                 yarnTask.yarnCommand = tokens
 
-                if (tokens.head().equalsIgnoreCase("run")) {
-                    yarnTask.dependsOn(YarnInstallTask.NAME)
+                if ( tokens.head().equalsIgnoreCase( "run" ) )
+                {
+                    yarnTask.dependsOn( YarnInstallTask.NAME )
                 }
 
                 return yarnTask